Api Security: /api/absence with OFFICE role and own data
